+++ R/xml_parse.R
@@ -98,6 +98,10 @@ read_xml.character <- function(x, encoding = "", ..., as_html = FALSE,
     stop("Document is empty", call. = FALSE)
   }
 
+  if (length(x) > 1) {
+    stop("`x` must be a string of length 1", call. = FALSE)
+  }
+
   options <- parse_options(options, xml_parse_options())
   if (grepl("<|>", x)) {
     read_xml.raw(charToRaw(enc2utf8(x)), "UTF-8", ..., as_html = as_html, options = options)